+++ example/example.dart
@@ -1,4 +1,11 @@
+import 'dart:js_interop';
+import 'package:web/web.dart';
 import 'package:poor_web/poor_web.dart';
+
+extension on Window {
+  @JS()
+  external Object? calc;
+}
 
 void main() {
   document.querySelector('#example')?.appendAll([
@@ -8,17 +15,14 @@ void main() {
 
   final global = globalThis;
 
-  global['calc'] = Object()..['add'] = Fn.a2((Number a, Number b) => a + b);
+  global.calc = Object()..['add'] = Fn.a2((Number a, Number b) => a + b);
 
   final a = 1;
   final b = 2;
-  final c = global['calc']?.object?['add']?.function?.call(
-    Number(a),
-    Number(b),
-  );
+  final c = global.calc?.object?['add']?.function?.call(Number(a), Number(b));
   print('$a + $b = $c'); // 1 + 2 = 3
 
-  global['calc']?.object?['print'] = Fn(
+  global.calc?.object?['print'] = Fn(
     ([a, b, c, _]) => switch ((a?.number, b?.number, c?.number)) {
       (Number a, Number b, Number c) => String('$a + $b = $c'),
       _ => null,