+++ lib/src/web.dart
@@ -1,10 +1,8 @@
 import 'dart:core' as core;
 import 'deps/web.dart' as web;
-export 'deps/web.dart';
+import 'deps/js_interop.dart';
 
-import 'js.dart';
-
-extension ElementAppendAll on web.Element {
+extension ElementAppendAllExtension on web.Element {
   void appendAll(core.Iterable<JSAny> elements) {
     for (final element in elements) {
       append(element);
@@ -12,17 +10,8 @@ extension ElementAppendAll on web.Element {
   }
 }
 
-extension type Window._(web.Window _window) implements web.Window, Object {
-  Any? operator [](core.String property) => (_window as Object)[property];
-
-  void operator []=(core.String property, Any? value) =>
-      (_window as Object)[property] = value;
-}
-
 /// Error on Node.js:
 ///   TypeError: null: type 'JSNull' is not a subtype of type 'JSObject'
-Window get window => Window._(globalContext as web.Window);
+web.Window get window => globalContext as web.Window;
 
-Window get globalThis => window;
-
-core.Uri get current => core.Uri.parse(window.location.href);
+web.Window get globalThis => window;